pyproject.toml
[tool.poetry]
name = "avito"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework-jsonapi = "^6.0.0"
djangorestframework-jwt = "^1.11.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
